master/requirements.txt
amqp==5.2.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
billiard==4.2.0
blinker==1.8.2
bootstrap4==0.1.0
Brotli==1.1.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
crispy-bootstrap4==2024.1
cron-descriptor==1.4.3
cryptography==43.0.0
cssselect2==0.7.0
decorator==4.4.2
diff-match-patch==20230430
Django==5.0.6
django-appconf==1.0.6
django-celery-beat==2.6.0
django-cors-headers==4.4.0
django-crispy-forms==2.2
django-crontab==0.7.1
django-dbbackup==4.1.0
django-extensions==3.2.3
django-filter==24.2
django-import-export==4.1.0
django-select2==8.2.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
Flask==3.0.3
fonttools==4.53.1
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.2
gspread==6.1.2
html5lib==1.1
httplib2==0.22.0
idna==3.7
imageio==2.34.2
imageio-ffmpeg==0.5.1
inflect==7.3.1
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
kombu==5.3.7
lxml==5.2.2
MarkupSafe==2.1.5
more-itertools==10.4.0
moviepy==1.0.3
mysqlclient==2.2.4
num2words==0.5.13
numpy==2.0.0
oauthlib==3.2.2
openpyxl==3.1.5
oscrypto==1.3.0
packaging==24.1
panda==0.3.1
pandas==2.2.2
pdfkit==1.0.0
pillow==10.4.0
proglog==0.1.10
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
PyMySQL==1.1.1
pyparsing==3.1.2
pypdf==4.3.1
pyphen==0.15.0
pypng==0.20220715.0
python-bidi==0.5.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==70.2.0
six==1.16.0
sqlparse==0.5.0
svglib==1.5.1
tablib==3.5.0
tinycss2==1.3.0
tqdm==4.66.4
typeguard==4.3.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
weasyprint==62.3
webencodings==0.5.1
Werkzeug==3.0.3
xhtml2pdf==0.2.16
xlrd==2.0.1
zopfli==0.2.3